refactor: DTO 클래스의 책임 분리 DTO클래스(CarNamesRequest.class) 에서 이루어지는 도메인 객체의 논리적인 특징에 대한 검증 로직 분리
