+++ src/main/java/racingcar/model/car/Car.java
@@ -1,8 +1,13 @@
 package racingcar.model.car;
 
+import java.util.regex.Pattern;
 import racingcar.model.car.strategy.MovingStrategy;
+import racingcar.util.ErrorMessage;
 
 public class Car {
+    private static final String VALID_WORD_REGEX = "(\\w)+";
+    private static final Pattern CAR_NAME_PATTERN = Pattern.compile(VALID_WORD_REGEX);
+    private static final int CAR_NAME_MAX_LENGTH = 5;
     private static final int DEFAULT_POSITION = 1;
 
     private final String carName;
@@ -10,6 +15,7 @@ public class Car {
     private final MovingStrategy movingStrategy;
 
     public Car(String carName, MovingStrategy movingStrategy) {
+        validate(carName);
         this.carName = carName;
         this.movingStrategy = movingStrategy;
     }
@@ -24,6 +30,23 @@ public class Car {
 
     public boolean movable() {
         return movingStrategy.movable();
+    }
+
+    private void validate(String carName) {
+        validateValidWord(carName);
+        validateOverMaxLength(carName);
+    }
+
+    private void validateValidWord(String splitCarName) {
+        if (!CAR_NAME_PATTERN.matcher(splitCarName).matches()) {
+            throw new IllegalArgumentException(ErrorMessage.INVALID_CAR_NAME_FORMAT.message());
+        }
+    }
+
+    private void validateOverMaxLength(String splitCarName) {
+        if (splitCarName.length() > CAR_NAME_MAX_LENGTH) {
+            throw new IllegalArgumentException(ErrorMessage.CAR_NAME_OVER_RANGE.message());
+        }
     }
 
     public String getCarName() {
